build: max-line-length check not catching all cases (#17570)
In tslint 5.19.0 they added a separate option that determines whether strings should be checked and it seems to have been defaulted to `false` (https://github.com/palantir/tslint/pull/4798). As a result we ended up with a few places where the limit was being violated.

These changes explicitly enable checking of strings and fix the failures that we had accumulated.