refactor(expansion): avoid unnecessary spacing check (#17693)
The check where `expanded` equals to `false`, and where the value of
the `_getExpandedState` method is compared to `default` will always
evaluate to `false`. This is because the method return value does not
have any overlap with `default`. Hence the check is unnecessary.

This seems to be a leftover from https://github.com/angular/components/commit/f9bd5d4bb53b71e40b89d51f77ea42da17a2c967.