chore: add FocusTrap demo page (#17858)
Note that much of this demo page is known to not be working with the
current focus trap implementation (ex. elements in the focus trap with
tabindex > 0 send focus out of the trap, elements in iframes/shadow DOM
get skipped when focus wraps, all focus traps can be escaped by clicking
on the URL and then pressing tab). The demo page will be helpful for
research into FocusTrap improvements (see
https://github.com/angular/components/issues/13054).