build: add target to create dev-app web package (#17937)
Creates a new Bazel target that can be built to create a
static web package of the dev-app. The web package can be
deployed to Firebase or other static file hosting services.

In a follow-up, we can create a script that replicates what
the old firebase deploy gulp task did.