refactor: fixes for theming api refactor and introduction of a theming API test
With the initial commits for the density system, a function for
retrieving the color configuration from the theme object has been
introduced.

This function is currently flawed since it always uses the theme object
as color configuration if the theme does not have a `color` property.
This is incorrect because the `color` config can be explicitly set to
`null`, or omitted if no color styles are desired.