fix(ng-update): better detection for workspace project in v9 hammer migration (#18525)
Currently the v9 HammerJS migration determines the workspace project by
consulting the `ts.Program` that has been constructed. This logic is not
guaranteed to work because a TypeScript program doesn't necessarily need
to contain any "root" file names. This means that we cannot reliably
determine the workspace project from the `ts.Program` and we will throw
an error that no project could be found.

We can improve this logic by simply using the workspace project that is
associated with the originating tsconfig file. This was previously not
possible, but 411d048249368e640cae4f31941dd67b24f07446 enabled us to
pass through the workspace project.

Fixes #18504.