refactor: use jasmine expectAsync instead of try/catch (#18535)
Currently tests that assert errors to be thrown/or not from
asynchronous methods use `try/catch` in order to use `await`.

We added a helper method for this in the private testing utilities. This
one reduced the `try/catch` burden. That helper method has downsides
though because it doesn't allow us to expect no error to be thrown.

We should remove this method in favor of using jasmine's `expectAsync`
function. This one gives us more flexibility and is also more
"jamsine"-idiomatic.