fix(google-maps): not rendering until mapTypeId is set (#18967)
In 2a6aae179a436886395cc8ecc631f7c7be4dff24 I added the ability to set the `mapTypeId` and after I was done with it and ran the unit tests, I added a default of `undefined` to the `DEFAULT_OPTIONS` so the tests didn't have to be updated. It looks like having the `undefined` in the defaults causes Google Maps not to render anything in the map until the map type changes. These changes remove the default value and update the tests instead.

Fixes #18965.