perf(ripple): optimize event registration (#18633)
* perf(ripple): optimize event registration

Improve scalabiliity of Angular Material components.
When used in an app with a large and heavy UI with large numbers of material components (for example in table cells), they can quickly become prohibitively expensive due to their heavy DOM structure and lots of up-front work.

Each component with a ripple ends up registering 6 event listeners. This PR makes 4 of them registered on-demand.

Note, that this is really a stop-gap solution to help improve performance in our application. A proper implementation should utilize a single document-level service that only registers one set of these event listeners, and uses event delegation to apply ripple effects.

* Update core.d.ts

* Add the @docs-private annotation.