refactor: cleanup fake event dispatch utilities (#19151)
Cleans up the fake event dispatch utiltities:

1. `dispatchMouseEvent` accepts an actual mouse event. This has been
introduced with [this commit](04bf3d1fd54402a5582f6043e2db02fde4598516).
If a mouse event has already been created, then it can be just
dispatched, but doesn't need to go through `dispatchMouseEvent` which
primarily is a shorthand for creating an event and dispatching it.

2. `dispatchEvent` helper should use a generic so that we don't
need to cast events unnecessarily.

3. `createMouseEvent` should not accept `x` and `y`. This is ambiguous.
It's not clear what coordinates are expected. To improve this,
parameters are now explicit, and a better comment has been added.