build: properly determine patch and minor branches for merge script
Currently we naively rely on the `package.json` version for determining
the minor and patch branches. This doesn't work out well, if the version
is set to a prerelease major.

This is because we'd need to determine the previous latest
patch branch from the Git upstream remote. This matches what framework
does. Only difference is that we are more strict, less reliant on Git,
and have machinery that prevents accidental versions from being picked up.

Not relying too much on Git helps as querying the `git ls-remote` all
the time could slow-down the caretaker from merging.

Note: Due to an upstream issue in ng-dev, we need to temporarily set
ts-node to use `commonjs`. by default, ts-node uses es2015 and we
cannot use that inside NodeJS.