build: add job that ensures view engine AOT compatibility (#19531)
Surprisingly with the switch to Ivy, we lost our infrastructure
that ensures View Engine AOT compatibility. View Engine compiler
is sometimes slighlty stricter in AOT as its pre-compiled factories
are in external classes. This requires variables to be "public" when
they are referenced in a template. In Ivy, these properties could
actually be `protected` or `private` too.

We already run tests against View Engine, but not in AOT. This is
because View Engine compiler does not process un-exported test classes.
In the future, with Ivy-only, we can run tests in AOT mode.