fix(material-experimental/mdc-form-field): do not show focus overlay when disabled
Fixes that the focus overlay is rendered/displayed when the form-field
is disabled. We fix this by just not rendering the focus overlay when
the form field is disabled. This should be better, compared to leaving
it in the DOM while toggling opacity or `display: none`, because it
saves bytes in SSR rendered apps, and helps reducing DOM nodes. The
assumption is that form-field's disabled state doesn't change constantly.