docs(stepper): error state example not set up correctly (#19416)
The `stepper-errors` example was set up using a linear stepper which doesn't allow users to move to a different step if the current one is invalid which means that the user doesn't have a way of seeing the error state.

These changes make the stepper non-linear, expand the example description to explain how the error state works, and add some text so users know what to do in order to see the error state.

Fixes #19407.